add test_file data path
