+++ MicrosoftActiveDirectory/tests/conftest.py
@@ -1,0 +1,16 @@
+from shutil import rmtree
+from tempfile import mkdtemp
+
+import pytest
+from sekoia_automation import constants
+
+
+@pytest.fixture
+def data_storage():
+    original_storage = constants.DATA_STORAGE
+    constants.DATA_STORAGE = mkdtemp()
+
+    yield constants.DATA_STORAGE
+
+    rmtree(constants.DATA_STORAGE)
+    constants.DATA_STORAGE = original_storage
